解决管理员处理学生选课容量冲突的问题

+++ Springboot/lab/src/main/java/com/example/lab/service/impl/CourseServiceImpl.java
@@ -173,11 +173,16 @@ public class CourseServiceImpl implements CourseService {
             return ResultMessage.NOTFOUND;
         }
         else {
-            try {
-                courseRepository.deleteById(courseId);
-                return ResultMessage.SUCCESS;
+            if (findCourseByCourseId(courseId).getStudents().isEmpty()){
+                try {
+                    courseRepository.deleteById(courseId);
+                    return ResultMessage.SUCCESS;
+                }
+                catch (Exception exception) {
+                    return ResultMessage.FAILED;
+                }
             }
-            catch (Exception exception) {
+            else {
                 return ResultMessage.FAILED;
             }
         }
